Refactor split toolbar - sidebar and pagination controls into separate widgets
