+++ packages/portal/src/components/media/MediaImageViewerControls.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     id="viewer-controls"
-    class="viewer-controls d-inline-flex align-items-center mx-auto"
+    class="viewer-controls position-absolute d-inline-flex align-items-center justify-content-center mx-auto"
   >
     <b-button
       v-b-tooltip.top="$t('media.controls.zoomIn')"
@@ -94,7 +94,17 @@
   @import '@europeana/style/scss/variables';
 
   .viewer-controls {
+    background-color: rgba($white, 0.95);
+    bottom: 0;
+    right: 0;
+    left: 0;
+    z-index: 3;
+    padding: 0.875rem 1rem;
+
     .btn {
+      background-color: transparent;
+      font-size: $font-size-large;
+
       &.disabled {
         opacity: 1;
         color: $middlegrey;
